// src/types/enums.ts
export enum PieceType {
    BLACK_PAWN = "bP",
    BLACK_ROOK = "bR",
    BLACK_KNIGHT = "bN",
    BLACK_BISHOP = "bB",
    BLACK_QUEEN = "bQ",
    BLACK_KING = "bK",
    WHITE_PAWN = "wP",
    WHITE_ROOK = "wR",
    WHITE_KNIGHT = "wN",
    WHITE_BISHOP = "wB",
    WHITE_QUEEN = "wQ",
    WHITE_KING = "wK",
    EMPTY = "--",
}

export enum PlayerColors {
    WHITE = "white",
    BLACK = "black",
}